requirements.txt
openai
python-dotenv
pyaudio
SpeechRecognition
pytesseract
mss
Pillow
